// iot-service/iot-angular/src/sdk/core/models/create-project-dto.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface CreateProjectDto {

  /**
   * Area type
   */
  areaType?: 'plant' | 'pipeline' | 'farm' | 'industrial' | 'other';

  /**
   * GeoJSON polygon for geofencing
   */
  geofence?: {
};

  /**
   * Owner ID (tenant)
   */
  idOwner: string;

  /**
   * Project name
   */
  name: string;

  /**
   * Project status
   */
  status?: string;
}